fastapi_app/requirements.txt
fastapi
uvicorn
pandas
numpy==1.23.5
scikit-learn
scikit-surprise
